# [OptimizeInstructions] Early return for functions with single childless body