It skips of optimize processing for some most frequent simple expressions like:
```wat
(func $0 (param $0) ... (nop))
(func $1 (param $0) ... (unreachable))
(func $2 (param $0) ... (return i32) (local.get $0))
(func $3 (param $0) ... (return i32) (i32.const 0))
```
Only in lit tests it may skip up to 8 functions and 2 in average per test. For real world examples it could be even more meaningful.