@MaxGraey 2021-09-25
Benchmark `wasm-opt -O3 -all-features` for 1.5mb real world wasm:
main branch:
```
15.82s user 0.64s system 877% cpu 1.875 total
15.73s user 0.58s system 884% cpu 1.845 total
15.70s user 0.61s system 895% cpu 1.823 total
```
this PR:
```
15.50s user 0.56s system 908% cpu 1.769 total
15.37s user 0.61s system 870% cpu 1.835 total
15.52s user 0.57s system 887% cpu 1.814 total
```