@kripken 2021-09-29
Hmm, that seems like a surprisingly big speedup if this just skips 27 functions in a 1.5MB file. How many functions are in there?

To try to reduce the chance of noise, how about measuring this with `BINARYEN_CORES=1` which will avoid nondeterminism with multiple threads?

Also, I'd suggest testing with `perf stat` on linux, like this:
```
$ perf stat wasm-opt [..]

 Performance counter stats for 'bin/wasm-opt waka.wasm -O3':

            417.56 msec task-clock                #    2.672 CPUs utilized          
               948      context-switches          #    0.002 M/sec                  
                71      cpu-migrations            #    0.170 K/sec                  
             3,116      page-faults               #    0.007 M/sec                  
     1,305,936,636      cycles                    #    3.128 GHz                    
     1,363,940,168      instructions              #    1.04  insn per cycle         
       302,349,897      branches                  #  724.087 M/sec                  
         7,332,592      branch-misses             #    2.43% of all branches        

       0.156258018 seconds time elapsed

       0.403095000 seconds user
       0.024933000 seconds sys
```
The key lines are cycles and instructions. Instructions should vary very little on a deterministic workload - almost not at all. Cycles will vary based on caching, but it can still be interesting.