@kripken 2021-09-30
I can try to find time to do detailed investigation here, but I worry it would be a waste of time. The speedup here is to skip 27 almost-empty functions out of a 1.5 MB wasm file. And that's in just one pass out of many that `-O3` does in your measurement - and one of the faster ones. Almost all the time - well over 99.99%, I'd bet - is spent outside of that one pass on those 27 almost-empty functions, so any speedup would be very very small.

Am I thinking about this wrong?