@MaxGraey 2021-09-30
If it can add 2-3% performance with just 3 lines of code, why not add it? Even if it doesn't work for all inputs.
It seems to me the perf gain is achieved by skipping many checks, avoiding call walker superclass, creating and initializing aux optimize walkers and etc