@kripken 2021-09-30
> If it can add 2-3% performance with just 3 lines of code, why not add it

I agree that any 2-3% speedup is worthwhile! But I don't see how it is possible for this to speed up that much:

* If this is a 2-3% speedup, then those 27 almost-empty functions took 2-3% of the time before.
* 27 almost-empty functions take something like 200 bytes.
* In a 1.5 MB module, those- 200 bytes are around much less (100x less) than 1%.

Perhaps just measuring size is inaccurate. We could also measure by the number of functions, which I see is over 6,000, so those 27 functions are much less than 2-3% of the total number of functions. And they are among the fastest functions to optimize in `-O3`, as we spend *lots* of time in complex functions with lots of locals. So I just don't see how this can be close to a noticeable speedup of even 2-3%.

But, that's all theory, so I found the time to measure this carefully using `perf`, and I see no real difference with the patch (something like `36,649,060,706` instructions with and without it, when using a single core to reduce noise).

But I am very much in favor of 2-3% speedups! Just we need to focus on things that matter more. One thing I noticed in #4165 is the large amount of allocations we do: using more SmallVector/SmallSet could be useful, for example.